% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otoliths_skj.R
\docType{data}
\name{otoliths_skj}
\alias{otoliths_skj}
\title{Otolith Data (Skipjack)}
\format{
Data frame containing two columns:
\tabular{ll}{
  \code{age} \tab age (years)\cr
  \code{len} \tab length (cm)
}
}
\source{
Macdonald, J., Day, J., Magnusson, A., Maunder, M., Aoki, Y., Matsubara, N.,
Tsuda, Y., McKechnie, S., Teears, T., Leroy, B., Castillo-Jordán, C.,
Hampton, J., and Hamer, P. (2022).
\emph{Review and new analyses of skipjack growth in the Western and Central
Pacific Ocean}.
Western and Central Pacific Fisheries Commission Report
WCPFC-SC18-2022/SA-IP-06.
\url{https://meetings.wcpfc.int/node/16254}.
}
\usage{
otoliths_skj
}
\description{
Simulated otolith data, loosely based on a skipjack tuna dataset analyzed by
Macdonald et al. (2022).
}
\examples{
otoliths_skj
}
\seealso{
\code{\link{otoliths_had}}, \code{otoliths_skj}, and \code{\link{tags_skj}}
are example datasets.

\code{\link{fimsgrowth-package}} gives an overview of the package.
}
