% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonbert.R
\name{vonbert}
\alias{vonbert}
\title{Von Bertalanffy Growth Model}
\usage{
vonbert(data, par, silent = TRUE, ...)
}
\arguments{
\item{data}{a data list.}

\item{par}{a parameter list.}

\item{silent}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{\dots}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{t}{age (vector).}

\item{L1}{predicted length at age \code{t1}.}

\item{L2}{predicted length at age \code{t2}.}

\item{k}{growth coefficient.}

\item{t1}{age where predicted length is \code{L1}.}

\item{t2}{age where predicted length is \code{L2}.}
}
\description{
Fit a von Bertalanffy growth model to otoliths and/or tags, using the
Schnute-Fournier parametrization.
}
